refactor(wallet): new mod storage with actor-independent code
