+++ wallet/src/storage/error.rs
